frontend: add bar for navigator
