+++ backend/frontend/src/components/Controller/index.js
@@ -1,19 +1,140 @@
 import React, { Component } from 'react'
 import { Route, Switch, Redirect } from 'react-router-dom'
 
-import { NAVIGATOR_URL, MENU_URL, SEARCH_URL } from '../../lib/consts'
+import { Toolbar, Typography } from 'material-ui'
+import {
+  faArrowAltCircleLeft,
+  faArrowAltCircleRight,
+  faBars,
+  faBookmark,
+  faHistory,
+  faList,
+  faSearch,
+  faSignOutAlt,
+  faWindowMinimize,
+} from '@fortawesome/fontawesome-free-solid'
+import { faSquare } from '@fortawesome/fontawesome-free-regular'
 
-import Controller from './Controller'
+import controller from '../../lib/controller'
+import {
+  BOOKMARKS_URL,
+  CONTROLLER_URL,
+  HISTORY_URL, MENU_URL,
+  NAVIGATOR_URL,
+  SEARCH_URL,
+} from '../../lib/consts'
+
+import ToolbarButton from './ToolbarButton'
 import Search from './Search'
 import Menu from './Menu'
-import Navigator from './Navigator'
+import Navigator, { Bar as NavigatorBar } from './Navigator'
 
 import './index.css'
 
 /**
+ * Renders the top navigation bar, showing the current path in the URL, and the hover state.
+ * @param title The title text in the top bar.
+ * @param history A `history` object.
+ * @param location A `location` object.
+ * @param onHover Fired on hover with name.
+ */
+const TopBar = ( { title, history, location, onHover } ) => {
+  const resetHover = () => onHover( null )
+
+  return (
+    <Toolbar className="top bar">
+      <ToolbarButton
+        name="Menu"
+        icon={faBars}
+        onClick={() => history.push( { ...location, pathname: MENU_URL } )}
+        onMouseEnter={() => onHover( 'Menu' )}
+        onMouseLeave={resetHover}
+      />
+      <ToolbarButton
+        name="Backwards"
+        icon={faArrowAltCircleLeft}
+        onClick={() => history.goBack()}
+        onMouseEnter={() => onHover( 'Backwards' )}
+        onMouseLeave={resetHover}
+      />
+      <ToolbarButton
+        name="Forwards"
+        icon={faArrowAltCircleRight}
+        onClick={() => history.goForward()}
+        onMouseEnter={() => onHover( 'Forwards' )}
+        onMouseLeave={resetHover}
+      />
+      <Typography className="name" type="title">{title}</Typography>
+      <ToolbarButton
+        name="Minimize"
+        icon={faWindowMinimize}
+        onClick={() => history.push( '/' )}
+        onMouseEnter={() => onHover( 'Minimize' )}
+        onMouseLeave={resetHover}
+      />
+      <ToolbarButton
+        name="Pop Out"
+        icon={faSignOutAlt}
+        onClick={() => window.open( `${CONTROLLER_URL}/?controllerOnly=true`, '_blank' )}
+        onMouseEnter={() => onHover( 'Pop Out' )}
+        onMouseLeave={resetHover}
+      />
+    </Toolbar>
+  )
+}
+
+/**
+ * Renders the bottom navigation bar.
+ * @param history A `history` object.
+ * @param renderContent A render prop for content in the bottom bar.
+ * @param location A `location` object.
+ * @param onHover Fired on hover with name.
+ */
+const BottomBar = ( { history, renderContent, location, onHover } ) => {
+  const go = pathname => () => history.push( { ...location, pathname } )
+  const resetHover = () => onHover( null )
+
+  return (
+    <Toolbar className="bottom bar">
+      <ToolbarButton name="Search" icon={faSearch} onClick={go( SEARCH_URL )} onHover={onHover} />
+      <ToolbarButton
+        name="Bookmarks"
+        icon={faBookmark}
+        onClick={go( BOOKMARKS_URL )}
+        onMouseEnter={() => onHover( 'Search' )}
+        onMouseLeave={resetHover}
+      />
+      <ToolbarButton
+        name="History"
+        icon={faHistory}
+        onClick={go( HISTORY_URL )}
+        onMouseEnter={() => onHover( 'History' )}
+        onMouseLeave={resetHover}
+      />
+      <div className="name">{renderContent()}</div>
+      <ToolbarButton
+        name="Navigator"
+        icon={faList}
+        onClick={go( NAVIGATOR_URL )}
+        onMouseEnter={() => onHover( 'Navigator' )}
+        onMouseLeave={resetHover}
+      />
+      <ToolbarButton
+        name="Clear"
+        icon={faSquare}
+        onClick={controller.clear}
+        onMouseEnter={() => onHover( 'Clear' )}
+        onMouseLeave={resetHover}
+      />
+    </Toolbar>
+  )
+}
+
+
+/**
  * Controller controls the display and configures settings.
  */
-class ControllerContainer extends Component {
+class Controller extends Component {
   constructor( props ) {
     super( props )
 
@@ -41,23 +162,33 @@ class ControllerContainer extends Component {
     const routes = [
       [ MENU_URL, Menu ],
       [ SEARCH_URL, Search ],
-      [ NAVIGATOR_URL, Navigator ],
+      [ NAVIGATOR_URL, Navigator, NavigatorBar ],
     ]
 
     return (
       <Switch>
-        {routes.map( ( [ route, Component ] ) => (
+        {routes.map( ( [ route, Component, BarComponent ] ) => (
           <Route
             key={route}
             path={route}
             render={props => (
-              <Controller
-                {...this.props}
-                {...props}
-                Component={Component}
-                onHover={this.onHover}
-                title={hovered || route.split( '/' ).pop()}
-              />
+              <div className="controller">
+                <TopBar
+                  {...props}
+                  title={hovered || route.split( '/' ).pop()}
+                  onHover={this.onHover}
+                />
+                <div className="content">
+                  <Component {...this.props} {...props} />
+                </div>
+                <BottomBar
+                  {...props}
+                  onHover={this.onHover}
+                  renderContent={() => BarComponent
+                    ? <BarComponent {...this.props} {...props} />
+                    : null}
+                />
+              </div>
             )}
           />
         ) )}
@@ -67,4 +198,4 @@ class ControllerContainer extends Component {
   }
 }
 
-export default ControllerContainer
+export default Controller
